+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - Issue [#237](https://github.com/DanielChappuis/reactphysics3d/issues/237) Wrong assert has been removed 
 - Issue [#239](https://github.com/DanielChappuis/reactphysics3d/issues/239) Memory allocation alignment
 - Issue [#240](https://github.com/DanielChappuis/reactphysics3d/issues/240) Uninitialized variable
+- Issue [#347](https://github.com/DanielChappuis/reactphysics3d/issues/347) Missing collision between capsule and triangle edge in some case
 - Issue with edge vs edge collision detection for BoxShape, ConvexMeshShape, ConcaveMeshShape and HeightFieldShape (SAT algorithm)
 - Compilation error on Clang 19